moved image exception class to a new 'exception' package
