+++ src/main/java/codes/pnk/model/domain/algorithm/Algorithm.java
@@ -1,7 +1,7 @@
 package codes.pnk.model.domain.algorithm;
 
 import codes.pnk.model.domain.common.Image;
-import codes.pnk.model.domain.common.ImageException;
+import codes.pnk.model.domain.exception.ImageException;
 import codes.pnk.model.domain.common.Text;
 
 public abstract class Algorithm {
