+++ src/main/java/codes/pnk/model/domain/algorithm/dh/lsb/LsbAlgorithm.java
@@ -2,7 +2,7 @@ package codes.pnk.model.domain.algorithm.dh.lsb;
 
 import codes.pnk.model.domain.algorithm.Algorithm;
 import codes.pnk.model.domain.common.Image;
-import codes.pnk.model.domain.common.ImageException;
+import codes.pnk.model.domain.exception.ImageException;
 import codes.pnk.model.domain.common.Text;
 
 import java.awt.image.BufferedImage;
